feat: abstract shared behaviour into ZerokitMerkleTree
