+++ utils/src/merkle_tree/mod.rs
@@ -1,3 +1,7 @@
+pub mod full_merkle_tree;
 #[allow(clippy::module_inception)]
 pub mod merkle_tree;
+pub mod optimal_merkle_tree;
+pub use self::full_merkle_tree::*;
 pub use self::merkle_tree::*;
+pub use self::optimal_merkle_tree::*;
